backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.6
alembic==1.10.3
websockets==11.0.2
pytest==7.3.1
httpx==0.23.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
